# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.BlockedTimeResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      end_time:
        type: string
      id:
        type: string
      is_recurring:
        type: boolean
      professional_id:
        type: string
      reason:
        type: string
      recurrence_rule:
        type: string
      start_time:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.CreateBlockedTimeRequest:
    properties:
      end_time:
        type: string
      is_recurring:
        type: boolean
      professional_id:
        type: string
      reason:
        maxLength: 255
        minLength: 3
        type: string
      recurrence_rule:
        type: string
      start_time:
        type: string
    required:
    - end_time
    - professional_id
    - reason
    - start_time
    type: object
...
  dto.ErrorResponse:
    properties:
      details:
        additionalProperties: true
        type: object
      error:
        type: string
      message:
        type: string
    type: object
...
  dto.ListBlockedTimesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.BlockedTimeResponse'
        type: array
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@nexo.barber
    name: API Support
    url: https://nexo.barber/support
  description: API completa para gestão de barbearias com módulos de Agendamento,
    Comanda, Financeiro, Estoque e Analytics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://nexo.barber/terms
  title: NEXO - Barber Analytics API
  version: "2.0"
paths:
...
  /api/v1/blocked-times:
    get:
      consumes:
      - application/json
      description: Lista bloqueios de horário com filtros opcionais
      parameters:
      - description: ID do profissional
        in: query
        name: professional_id
        type: string
      - description: Data inicial (RFC3339)
        in: query
        name: start_date
        type: string
      - description: Data final (RFC3339)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListBlockedTimesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar bloqueios de horário
      tags:
      - Bloqueios
    post:
      consumes:
      - application/json
      description: Cria um novo bloqueio de horário na agenda
      parameters:
      - description: Dados do bloqueio
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBlockedTimeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.BlockedTimeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflito com bloqueio existente
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar bloqueio de horário
      tags:
      - Bloqueios
  /api/v1/blocked-times/{id}:
    delete:
      consumes:
      - application/json
      description: Remove um bloqueio de horário
      parameters:
      - description: ID do bloqueio
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deletar bloqueio de horário
      tags:
      - Bloqueios
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token JWT (formato: "Bearer {token}")'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...